Removed test util class, methods moved to `ListTestUtil`
